@@ -1,3 +1,3 @@
 module github.com/jub0bs/namecheck
 
-go 1.21
+go 1.23.3
